Update namespaces for servers, databases.configured, and databases.setup
