+++ mathesar_ui/src/api/rpc/index.ts
@@ -5,7 +5,6 @@ import { buildRpcApi } from '@mathesar/packages/json-rpc-client-builder';
 import { collaborators } from './collaborators';
 import { columns } from './columns';
 import { constraints } from './constraints';
-import { database_setup } from './database_setup';
 import { databases } from './databases';
 import { records } from './records';
 import { roles } from './roles';
@@ -19,7 +18,6 @@ export const api = buildRpcApi({
   getHeaders: () => ({ 'X-CSRFToken': Cookies.get('csrftoken') }),
   methodTree: {
     collaborators,
-    database_setup,
     databases,
     records,
     roles,
